SB-902: Some Groovy IT improvements
